+++ src/app/App.js
@@ -5,15 +5,13 @@ import Users from "./layouts/users";
 import Login from "./layouts/login";
 import Main from "./layouts/main";
 import NavBar from "./components/ui/navBar";
-import EditUserPage from "./components/page/editUserPage/editUserPage";
 
 function App() {
     return (
         <div>
             <NavBar />
             <Switch>
-                <Route path="/users/:userId?/edit" component={EditUserPage} />
-                <Route path="/users/:userId?" component={Users} />
+                <Route path="/users/:userId?/:edit?" component={Users} />
                 <Route path="/login/:type?" component={Login} />
                 <Route path="/" exact component={Main} />
                 <Redirect to="/" />
