+++ src/app/components/page/editUserPage/index.js
@@ -1,0 +1,2 @@
+import EditUserPage from "./editUserPage";
+export default EditUserPage;
