+++ src/app/layouts/users.jsx
@@ -3,11 +3,24 @@ import { useParams } from "react-router-dom";
 
 import UserPage from "../components/page/userPage";
 import UsersListPage from "../components/page/userListPage";
+import EditUserPage from "../components/page/editUserPage/editUserPage";
 
 const Users = () => {
     const params = useParams();
-    const { userId } = params;
-    return <>{userId ? <UserPage userId={userId} /> : <UsersListPage />}</>;
+    const { userId, edit } = params;
+    return (
+        <>
+            {userId ? (
+                edit ? (
+                    <EditUserPage />
+                ) : (
+                    <UserPage userId={userId} />
+                )
+            ) : (
+                <UsersListPage />
+            )}
+        </>
+    );
 };
 
 export default Users;
